making read functions use plpgsql so they can use if statement to perform permission check BEFORE attempting the delete.  This allows the error to be thrown for invalid id/users, even if they do not exist.  Making the functions explicitly RETURN QUERY (before the select statement) because they are now plpgsql functions.
Adding readonly tables for farm_permissions and poultry_inspections so they can be used as return types for the read() functions and have parameter names (id) similar to the function inputs, for plpgsql functions.  This simplifies the functions, but increases the number of tables, even though these are never inserted into.